backend/requirements.txt
Flask
pymongo
flask-cors
scikit-learn
python-dotenv